requirements.txt
python-binance>=1.0.20
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
python-socks[asyncio]>=2.4.4
